Seperate token validator
